testing/web-platform/tests/shadow-dom/declarative/declarative-shadow-dom-available-to-element-internals.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Declarative Shadow DOM</title>
<link rel="author" href="mailto:avandolder@mozilla.com">
<link rel="help" href="https://html.spec.whatwg.org/#parsing-main-inhead:available-to-element-internals">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<custom-element>
    <template shadowrootmode="open">
    </template>
</custom-element>

<script>
    let internals;
    customElements.define("custom-element", class extends HTMLElement {
        constructor() {
            super();
            internals = this.attachInternals();
        }
    });

    window.onload = () => {
        test(() => {
            assert_true(!!internals);
            assert_true(!!internals.shadowRoot);
            assert_equals(internals.shadowRoot, document.querySelector("custom-element").shadowRoot);
        }, "Declarative Shadow DOM: shadow root should be available to element internals");
    };
</script>
